Look. Refactors. Move more functionality to base class.
